pyproject.toml
[tool.poetry]
name = "reqres"
version = "0.1.0"
description = ""
authors = ["Maksim Kudaev <kudaev.m@nakukop.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.1"
selene = "2.0.0b16"
dynaconf = "^3.1.11"
requests = "^2.28.2"
pytest-voluptuous = "^1.2.0"
voluptuous = "^0.13.1"
allure-pytest = "^2.12.0"
curlify = "^2.2.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = 1
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = " -vv -s --alluredir=allure-results --allure-no-capture "
